weather-analysis/pyproject.toml
[project]
name = "weather-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jupyterlab>=4.5.1",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
]
